+++ cuda/src/reading_data/reading_data.cuh
@@ -13,9 +13,9 @@
 
 #define MAX_LINE_LENGTH 1024
 
-__host__ unsigned int get_num_nodes(const char* filename);
+__host__ size_t get_num_nodes(const char* filename);
 __host__ int get_num_values(const char* line);
 __host__ boolean getV(const char* filename, Node* v);
-__host__ boolean getA(const char* filename, const Node* v, lex_product* a, unsigned int n);
+__host__ boolean getA(const char* filename, const Node* v, lex_product* a, size_t n);
 
 #endif
