+++ cuda/src/utilities/utilities.cuh
@@ -2,9 +2,11 @@
 #ifndef UTILITIES 
 #define UTILITIES
 
-typedef unsigned int node;
+#include <stddef.h>
+
+typedef size_t node;
 typedef short int boolean;
 
-__host__ __device__ unsigned int idx(node i, node j, unsigned int n);
+__host__ __device__ size_t idx(node i, node j, size_t n);
 
 #endif
